Typedef c_clock_policy::duration_type to long double
Reason: double would lose precision with very high (64-bit) std::clock_t
values. Note that this also removes some -Wconversion warnings from GCC
(Clang, instead, didn't warn).

+++ breath/time/c_clock_policy.hpp
@@ -26,7 +26,7 @@ namespace breath {
 class c_clock_policy
 {
 public:
-    typedef double      duration_type ;
+    typedef long double duration_type ;
 
                         c_clock_policy() ;
